chore(structure): rename common to shared
