+++ apps/frontend/src/features/categories/categories.api.ts
@@ -3,7 +3,7 @@ import type {
   CategoryWithComputed,
   Paginated,
 } from "@recipes/shared";
-import { apiClient } from "@/common/api/client";
+import { apiClient } from "@/shared/api/client";
 
 export function getCategories(
   filters: Partial<CategoryQuery> = {},
