+++ apps/frontend/src/shared/ui/AppLogo.vue
