+++ apps/frontend/src/shared/ui/Footer.vue
