+++ apps/frontend/src/shared/ui/Header.vue
