+++ apps/frontend/src/shared/ui/Rating.vue
