+++ apps/frontend/src/shared/ui/Section.vue
