+++ apps/frontend/src/shared/ui/SectionHeader.vue
