+++ apps/frontend/src/shared/ui/WavyDivider.vue
