+++ apps/frontend/src/shared/ui/_footer/FooterCopyright.vue
