+++ apps/frontend/src/shared/ui/_footer/FooterNav.vue
