+++ apps/frontend/src/shared/ui/_footer/FooterSocials.vue
