+++ apps/frontend/src/shared/ui/_header/HeaderUser.vue
